requirements.txt
fal_client
requests
flask
flask-cors
pillow 